cleanup & reorganise src folder
