+++ src/components/dashboard/DashboardConflicts.vue
