+++ src/components/shifts/ShiftBulkActions.vue
@@ -48,9 +48,9 @@
 </template>
 
 <script>
-import ShiftAssignContractDialog from "@/components/ShiftAssignContractDialog";
-import ShiftBulkActionsDialogDelete from "@/components/ShiftBulkActionsDialogDelete";
-import ShiftBulkActionsDialogReview from "@/components/ShiftBulkActionsDialogReview";
+import ShiftAssignContractDialog from "@/components/shifts/ShiftAssignContractDialog";
+import ShiftBulkActionsDialogDelete from "@/components/shifts/ShiftBulkActionsDialogDelete";
+import ShiftBulkActionsDialogReview from "@/components/shifts/ShiftBulkActionsDialogReview";
 import {
   mdiCheck,
   mdiCheckAll,
