+++ src/components/shifts/ShiftBulkActionsDialogDelete.vue
